feature: reorganize folder structure and componentize some components
